Change AP wording
I think that the description for AP should be changed. It currently says that the most recent version of the data might not be the latest. 
The explanation of Consistency uses the same word, recent. 
I have changed AP to qualify 'recent' as being the most recent data available on the node